Hotfix: money response schemas must be float, not int (unblocks signups)
After the pennies migration, per-minute DEBITs are fractional (£5.20) and
balances hold pennies (£4.60). The Pydantic response schemas still typed these
as int, so Pydantic raised int_from_float and 500'd every read:
/transactions/me/balance, /transactions/me, /admin/transactions/all, and the
admin user list/detail. The client then showed balance 0 — which looked like
"new signups get no welcome credit", but the DB was correct all along (user 35:
BONUS +15, two DEBIT 5.20, balance 4.60 — the per-minute billing works).

Changed to float: TransactionOut.amount/balance_before/balance_after,
BalanceResponse.balance/credit_balance/paid_balance (+ last_updated Optional),
Refund amount/refunded_amount, AdminUserListItem/AdminUserDetail.balance, and
the admin gift/adjust/create/update balance request fields. Backend-only, no
migration.

+++ TAROT-BACKEND/app/schemas/user.py
@@ -80,7 +80,7 @@ class AdminUserListItem(BaseModel):
     email: str
     role: Role
     status: UserStatus
-    balance: int
+    balance: float  # money stored to 2 dp (pennies)
     is_verified: bool
     is_online: bool
     created_at: datetime
@@ -98,7 +98,7 @@ class AdminUserDetail(BaseModel):
     email: str
     role: Role
     status: UserStatus
-    balance: int
+    balance: float  # money stored to 2 dp (pennies)
     is_verified: bool
     is_online: bool
     price_per_second: Optional[float] = None
@@ -120,7 +120,7 @@ class AdminUserCreate(BaseModel):
     role: Role = Role.USER
     status: UserStatus = UserStatus.ACTIVE
     is_verified: bool = False
-    balance: int = 0
+    balance: float = 0
     price_per_second: Optional[float] = None
     bio: Optional[str] = None
 
@@ -144,7 +144,7 @@ class AdminUserUpdate(BaseModel):
     price_per_second: Optional[float] = None
     bio: Optional[str] = None
     password: Optional[str] = None
-    balance: Optional[int] = None
+    balance: Optional[float] = None
 
     @field_validator("password")
     @classmethod
@@ -168,13 +168,13 @@ class UserRoleUpdate(BaseModel):
 
 class AdminGiftBalance(BaseModel):
     """Schema for admin gift balance"""
-    amount: int = Field(gt=0, description="Amount of points to gift")
+    amount: float = Field(gt=0, description="Amount of points to gift")
     message: str = Field(default="", max_length=500, description="Optional gift message")
 
 
 class AdminBalanceAdjustment(BaseModel):
     """Schema for admin balance adjustment"""
-    amount: int
+    amount: float
     reason: str
 
     @field_validator("amount")
